feat: SQL frontend (Phase 6) + Q5/Q10 multi-join + groupby int64 fix
sql_frontend.py — new module: sqlglot → LogicalPlan translator
  - SELECT/FROM/WHERE/GROUP BY/ORDER BY/LIMIT
  - INNER/LEFT JOIN with schema-aware key assignment (handles col alias remapping
    when right join keys are dropped from output)
  - BETWEEN, IN/NOT IN, LIKE (prefix%), CASE WHEN ... END
  - All aggregates: SUM AVG MIN MAX COUNT(*)
  - Arithmetic: + - * / with float coercion for integer literals in arith context
  - Chained multi-table JOINs with automatic left/right key detection
  - mx.sql() public API exported from __init__.py

custom_ops.py — fix _build_group_ids composite key to use int64 instead of
  int32 — prevents OverflowError with many-column groupby (Q10-style)

Tests:
  scripts/_test_phase_sql.py — 8 tests: simple filter/sum/sort, Q6 Q1 Q3 Q12 Q14
    all produce correct results vs DuckDB
  scripts/_test_phase_q5_q10.py — Q5 (6-way join) and Q10 (4-way join + limit)
    validate multi-table join correctness vs DuckDB

pixi.toml — add test-sql, test-q5q10 tasks; include both in test-all

+++ custom_ops.py
@@ -1152,10 +1152,11 @@ class CustomOpsCompiler(GraphCompiler):
                 s *= sizes[j]
             strides.append(s)
 
-        composite = np.zeros(len(table), dtype=np.int32)
+        # Use int64 to avoid overflow when strides are large (many-column groupby)
+        composite = np.zeros(len(table), dtype=np.int64)
         for i, enc in enumerate(encodings):
-            idx = enc.indices.to_numpy(zero_copy_only=False)
-            composite += (idx * strides[i]).astype(np.int32)
+            idx = enc.indices.to_numpy(zero_copy_only=False).astype(np.int64)
+            composite += idx * np.int64(strides[i])
         return composite
 
     def _hash_join_mojo_cpu(self, left_keys, right_keys):
